fix: the 127.0.0.1 virtualhost is before the inline:80 virtualhost
This was made because if no inline interface is present it's vhost will be 127.0.0.1 and before this change this was put before the openvas callback vhost making it unreachable. As a quick fix, I re-ordered the vhosts but it issues a warning. This should be fixed for good in the next release when we will support DNS DNAT'ing in inline mode

Monotone-Parent: d448e7faeecfdacc230415ba8fc11906c69831f4
Monotone-Revision: 4e6db62ded579faf798cf45ee4a0b4de3ae38e96

Monotone-Author: obilodeau@inverse.ca
Monotone-Date: 2012-02-22T20:45:24
Monotone-Branch: org.packetfence.trunk