bugfix: named not starting if vlan mode is not enabled
This is the second stab at the issue. This fix gets rid of the cruft of having to modify two locations when you want to alter the services list which will prevent issues like this one from re-occuring.

The actual fix could be described as: unifying the 'service that should be run' list in one location: service_list.