introduced loadbalanced_combined LogFormat
Makes it easier for people to have client IP in apache logs. I decided
against integrating it by default because it breaks from the well-known
combined (and well parsed) LogFormat. I think this is a good compromise
considering the corner use-case and the fact that I documented it in the
parameter.