Revamp Interface controller + admin integration
The Interface controller/model is now used in both the configurator and
the admin. The design has changed to handle the requirements of the
admin: the interface type is specified in the interface modal editor and
the validation of all types is limited to the configurator.

The common JavaScript has been structured in model/view "classes" for
easier maintenance (see js/interfaces.js).