Fixed a scoping issue with the regexps.
The regexps have to be defined inside the sub otherwise they won't yet
have been assigned by the time the sub is called.
Placing them at the top of the file or before the call to
$parse_alert_for would also do it but would break the locality of the
code. We could do it if performance turns out to be an issue as that
would avoid recompiling the regexp every time the sub is called.