Use field specific types for form data
These are useful for mobile devices, as they bring up the correct keyboard for data entry.

Doing this also has the useful effect of helping to clean the form data before it is submitted, and cause less user errors. 
Auto filled data from a device's internal store tends to include a space at the end. On iOS and OSX, this is cleaned by the OS before submittal.