remove 'This is bad' from pfconfig error on L2 write
Since it may or may not be bad if you're in active active with haproxy as db balancer with cluster enabled and if its 15 to 3 on a Wednesday afternoon, seeing 'This is bad' may cause insecurity even to the most experienced power users when it is not bad in the case defined previously. Having users happiness as a first priority, we shall not make them insecure even if the error message is meaningful 90% of the time.

Note: tried to my best to display the content of a popular 4 letters website instead of 'this is bad' in order to increase even more user happiness, but have failed, as to my deception, SSH was not able to display it with the appropriate level of detail.