add requested host in haproxy's portal config file
Add additional logging in haproxy that shows the "Host" header in haproxy's log files. This can greatly help identify missing "passthrough exceptions"  for the walled garden or needed for Google/Facebook/etc OAuth logins as we can see both the URL and HOST of the request that hit the portal instead of the passthrough.