Add new POTD portal auth module (WIP)
It's a copy of Authentication::Password but without a fixed
username (it uses the one provided by the user just like
Authentication::Login). The difference is that PID and username
are treated independently. Username is only considered for
authentication while PID is used to actually create the user
in DB.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>