add support for tagged voice VLAN on Aruba switches
Egress-VLAN-Name is used on HP::SWitch template
but it looks like it doesn't work on recent switches.
I don't want to break anything so I edited this switch module.

This attribute works on Aruba 2530 running Aruba OS 16.11
voice role needs to be defined and match **name** of the voice VLAN on the switch.