usb-bootable-iso: Fix GRUB installation to show device selection menu
The previous configuration was asking for manual /dev entry instead of
showing a list of devices. This was caused by only_debian=true which
tries to auto-detect but fails with multiple disks.

Changes:
- Set grub-installer/only_debian to false (forces device selection menu)
- Add grub-installer/with_other_os boolean true (allows installation with other OS)
- Keep bootdev unset so user is prompted with a list of devices
- Add temporary doc files to .gitignore

This should now show a proper device selection menu and prevent GRUB
from being installed to the USB stick instead of the target disk.